move ilrs and space weather to dedicated modules, ignore legacy data module in stub generation
