--- src/tudatpy/data_input/environment_data/ilrs.py
@@ -1,27 +1,0 @@
-"""ILRS SINEX station data loading."""
-
-from tudatpy.kernel.data_input.environment_data.ilrs import (
-    IlrsStationRegistryEntry,
-    SinexStationEccentricity,
-    SinexStationState,
-    convert_sinex_datetime_to_seconds_since_epoch,
-    read_domes_id_numbers,
-    read_ground_station_names,
-    read_ilrs_station_registry_from_sinex_site_id,
-    read_monument_numbers,
-    read_sinex_station_data,
-    read_sinex_station_eccentricities,
-)
-
-__all__ = [
-    "IlrsStationRegistryEntry",
-    "SinexStationEccentricity",
-    "SinexStationState",
-    "convert_sinex_datetime_to_seconds_since_epoch",
-    "read_domes_id_numbers",
-    "read_ground_station_names",
-    "read_ilrs_station_registry_from_sinex_site_id",
-    "read_monument_numbers",
-    "read_sinex_station_data",
-    "read_sinex_station_eccentricities",
-]
