--- src/tudatpy/data_input/environment_data/space_weather.py
@@ -1,13 +1,0 @@
-"""Space-weather data loading."""
-
-from tudatpy.kernel.data_input.environment_data.space_weather import (
-    SolarActivityContainer,
-    SolarActivityData,
-    read_solar_activity_data,
-)
-
-__all__ = [
-    "SolarActivityContainer",
-    "SolarActivityData",
-    "read_solar_activity_data",
-]
