+++ src/tudatpy/data_input/environment_data/space_weather/__init__.py
@@ -1,0 +1,1 @@
+from tudatpy.kernel.data_input.environment_data.space_weather import *
