+++ src/tudatpy/data_input/environment_data/space_weather/expose_space_weather.h
